refactor: rename middleware directory name
